requirements.txt
reportlab>=4.0.0
Pillow>=10.0.0